security-manager-policy-reload: don't print errors on image build
The policy reload script, when recreating user type buckets for Cynara,
first tries to erase them and then create fresh ones. But on the first run,
during image build, there are no user type buckets in Cynara.

The error during erase is ignored by the script, as it should be, but misleading
error messages are also printed to stderr.
The error messagess should be silenced.

Change-Id: Ib09651560f15263793d758698e792a01471e1657