Fix doxygen comments
This fixes all of the doxygen comment tags (/**) and a plethora of other
errors and inconsistencies. Mostly missing argument names for \param.

This is not a comprehensive doxygen comments review, but it does fix all
the doxygen errors and warnings and makes it possible to generate
somewhat correct doxygen documentation.

Change-Id: Ib030dab7a5c116a8a6a9ccb3665dd79163a7b632